mars_photos_app_backend/cloud_function/requirements.txt
Flask==2.0.1
google-cloud-storage==1.39.0
google-cloud-firestore==2.1.3